requirements.txt
python-dotenv==0.17.0
discord.py==1.7.0
discord-py-slash-command=1.1.1